ai/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
tensorboard>=2.14.0
requests>=2.31.0
pydantic>=2.4.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
